remove layout from ClientListScreen to ClientListTable
