+++ app/Orchid/Screens/Client/ClientListScreen.php
@@ -3,6 +3,7 @@
 namespace App\Orchid\Screens\Client;
 
 use App\Models\Client;
+use App\Orchid\Layouts\Client\ClientListTable;
 use Carbon\Carbon;
 use Carbon\CarbonPeriod;
 use Orchid\Screen\Screen;
@@ -56,23 +57,7 @@ class ClientListScreen extends Screen
     public function layout(): array
     {
         return [
-            Layout::table('clients', [
-                TD::make('phone', 'Телефон')->width('150px')->cantHide()->canSee($this->isWorkTime())->filter(TD::FILTER_TEXT),
-                TD::make('status', 'Статус')->render(function (Client $client) {
-                    return $client->status === 'interviewed' ? 'Опрошен' : 'Не опрошен';
-                })->width('150px')->popover('Статус по результатам работы оператора')->sort(),
-                TD::make('email', 'Email'),
-                TD::make('assessment', 'Оценка')->width('200px')->align(TD::ALIGN_RIGHT),
-                TD::make('created_at', 'Дата создания')->defaultHidden(),
-                TD::make('updated_at', 'Дата обновления')->defaultHidden(),
-            ])
+            ClientListTable::class
         ];
-    }
-
-    // Пример использования для метода -> canSee, который зависит от логического значения
-    private function isWorkTime():bool
-    {
-        $lunch = CarbonPeriod::create('13:00', '14:00');
-        return $lunch->contains(Carbon::now(config('app.timezone'))) === false;
     }
 }
